+++ main/include/eudaq/Time.hh
@@ -28,7 +28,7 @@ namespace eudaq {
 
 
       
-      explicit Time(long sec, long usec = 0) {
+      explicit Time(int32_t sec, int32_t usec = 0) {
         tv_usec = usec % 1000000;
         tv_sec = sec + usec / 1000000;
       }
@@ -74,8 +74,8 @@ namespace eudaq {
       std::string Formatted(const std::string & format = TIME_DEFAULT_FORMAT) const;
       static Time Current();
     private:
-      long tv_sec;
-      long tv_usec;
+      int32_t tv_sec;
+      int32_t tv_usec;
   };
 
   inline Time operator + (const timeval & lhs, const timeval rhs) {
